update tile slicing and overlaying logic
* use only one directory for tiles
* pass extra argument matrix to pipeline and to inference container

+++ am/segment/preprocess.py
@@ -13,21 +13,19 @@ from am.utils import clean_dir, read_image
 logger = logging.getLogger('am-segm')
 
 
-def slice_to_tiles(input_data_path, overwrite=False):
+def slice_to_tiles(input_path, output_path, overwrite=False):
     logger.info('Converting images to tiles')
 
-    tiles_path = input_data_path.parent / (input_data_path.stem + '_tiles')
-    if tiles_path.exists():
+    if output_path.exists():
         if overwrite:
-            clean_dir(tiles_path)
+            clean_dir(output_path)
         else:
-            logger.info('Tiles already exist. Skipping')
+            logger.info(f'{output_path} path already exists. Skipping')
             return
-
-    tiles_path.mkdir(parents=True, exist_ok=True)
+    output_path.mkdir(parents=True, exist_ok=True)
 
     image_paths = []
-    for root, dirs, files in os.walk(input_data_path):
+    for root, dirs, files in os.walk(input_path):
         if not dirs:
             for f in files:
                 image_paths.append(Path(root) / f)
@@ -46,14 +44,9 @@ def slice_to_tiles(input_data_path, overwrite=False):
             factor = max_size / max(image.shape)
             image = cv2.resize(image, None, fx=factor, fy=factor, interpolation=cv2.INTER_AREA)
 
-        image_tiles_path = tiles_path / image_path.parent.name / image_path.stem
-        if image_tiles_path.exists():
-            if overwrite:
-                clean_dir(image_tiles_path)
-            else:
-                logger.debug(f'Already exists: {image_tiles_path}')
-        else:
-            image_tiles_path.mkdir(parents=True)
+        group = image_path.parent.name
+        image_tiles_path = output_path / group / image_path.stem
+        image_tiles_path.mkdir(parents=True, exist_ok=True)
 
         tile_row_n, tile_col_n = compute_tile_row_col_n(image.shape, tile_size)
         target_size = (tile_row_n * tile_size, tile_col_n * tile_size)
@@ -62,8 +55,7 @@ def slice_to_tiles(input_data_path, overwrite=False):
         h, w = map(int, image.shape)
         meta['image'] = {'h': h, 'w': w}
         meta['tile'] = {'rows': tile_row_n, 'cols': tile_col_n, 'size': tile_size}
-        group_path = image_tiles_path.parent
-        json.dump(meta, open(group_path / 'meta.json', 'w'))
+        json.dump(meta, open(output_path / group / 'meta.json', 'w'))
 
         for i, tile in enumerate(tiles):
             tile_path = image_tiles_path / f'{i:03}.png'
@@ -79,15 +71,15 @@ def stitch_and_crop_tiles(tiles_path, tile_size, meta):
     tiles = [None] * len(tile_paths)
     for path in tile_paths:
         i = int(path.stem)
-        tiles[i] = cv2.imread(str(path))[:,:,0]  # because ch0==ch1==ch2
+        tiles[i] = cv2.imread(str(path))[:, :, 0]  # because ch0==ch1==ch2
 
     stitched_image = stitch_tiles(tiles, tile_size, meta['tile']['rows'], meta['tile']['cols'])
     stitched_image = CenterCrop(meta['image']['h'], meta['image']['w']).apply(stitched_image)
     return stitched_image
 
 
-def stitch_tiles_at_path(input_path, meta_path, overwrite=False, image_ext='png'):
-    output_path = Path(str(input_path) + '_stitched')
+def stitch_tiles_at_path(input_path, overwrite=False, image_ext='png'):
+    output_path = input_path.parent / 'tiles_stitched'
     if overwrite:
         rmtree(output_path, ignore_errors=True)
     output_path.mkdir(exist_ok=True)
@@ -96,15 +88,16 @@ def stitch_tiles_at_path(input_path, meta_path, overwrite=False, image_ext='png'
         logger.info(f'Stitching tiles at {group_path}')
         group = group_path.name
 
-        meta = json.load(open(meta_path / group / 'meta.json'))
+        meta = json.load(open(input_path / group / 'meta.json'))
         for image_type in ['source', 'mask']:
-            stitched_image = stitch_and_crop_tiles(group_path / image_type, 512, meta)
-            if stitched_image.max() <= 1:
-                stitched_image *= 255
+            if (group_path / image_type).exists():
+                stitched_image = stitch_and_crop_tiles(group_path / image_type, 512, meta)
+                if stitched_image.max() <= 1:
+                    stitched_image *= 255
 
-            stitched_group_path = output_path / group
-            stitched_group_path.mkdir(exist_ok=True)
+                stitched_group_path = output_path / group
+                stitched_group_path.mkdir(exist_ok=True)
 
-            stitched_image_path = stitched_group_path / f'{image_type}.{image_ext}'
-            cv2.imwrite(str(stitched_image_path), stitched_image)
-            logger.info(f'Saved stitched image to {stitched_image_path}')
+                stitched_image_path = stitched_group_path / f'{image_type}.{image_ext}'
+                cv2.imwrite(str(stitched_image_path), stitched_image)
+                logger.info(f'Saved stitched image to {stitched_image_path}')
